+++ frontend-new/app/components/file-preview/utils.ts
@@ -153,6 +153,44 @@ export function getRendererType(mimeType: string, fileName: string): string {
 }
 
 
+/** MIME types for PowerPoint presentation files (.ppt, .pptx) */
+export const PPT_MIME_TYPES = [
+  'application/vnd.ms-powerpoint',
+  'application/vnd.openxmlformats-officedocument.presentationml.presentation',
+];
+
+/** OOXML Word MIME type (.docx). */
+export const DOCX_MIME_TYPE =
+  'application/vnd.openxmlformats-officedocument.wordprocessingml.document';
+
+/**
+ * Checks whether a file is a PowerPoint presentation (PPT/PPTX) by MIME type or extension.
+ * PPT/PPTX files require server-side conversion to PDF via the `convertTo=pdf` query param
+ * on the streaming API before they can be previewed in the browser.
+ */
+export function isPresentationFile(mimeType?: string, fileName?: string): boolean {
+  if (mimeType && PPT_MIME_TYPES.includes(mimeType)) return true;
+  if (fileName) {
+    const ext = fileName.split('.').pop()?.toLowerCase();
+    if (ext === 'ppt' || ext === 'pptx') return true;
+  }
+  return false;
+}
+
+/**
+ * Checks whether a file is an OOXML Word doc (.docx). DOCX is rendered client-side
+ * by `docx-preview` directly from the in-memory Blob, so we skip `createObjectURL`
+ * and the extra `fetch` round-trip the DOCX renderer would otherwise perform.
+ */
+export function isDocxFile(mimeType?: string, fileName?: string): boolean {
+  if (mimeType === DOCX_MIME_TYPE) return true;
+  if (fileName) {
+    const ext = fileName.split('.').pop()?.toLowerCase();
+    if (ext === 'docx') return true;
+  }
+  return false;
+}
+
 /**
  * Determine if pagination controls should be visible
  */
